Add pagination support
I put the request parameters in the state, so the results can load them, and submit them with page++ for the next page.

I’ll need to include the results size in the query response as to more effectively paginate, i.e disabling the load more if there’s no more results.