Remove errors upon successful email sign in
I didn’t realise that the AUTHENTICATE_ actions were just do with attempting to authenticate from persisted session. When the user successfully signs in, the action is EMAIL_, and as such, after the user signed in, the error was still in the state. It will now clear with successful sign in.