Configure redux auth before rendering
It took me fucking ages to realise why a promise was used. It was specifically to avoid rendering the app while auth was still being deserialised which puts it in a weird transient state. Now that I configure it before rendering, I can rely upon the isLoggedIn attribute of redux auth.